EXT2,reportes,usrs,files

+++ frontend/backend/core/commands/fdisk.h
@@ -1,14 +1,13 @@
 #ifndef FDISK_H
 #define FDISK_H
-
 #include <string>
+#include <cstdio>
 
 class FDisk {
 public:
-    bool existeNombreLogica(FILE *disk, int startExtendida, std::string nombre);
-    void printLogicas(FILE *disk, int startExtendida);
+    bool existeNombreLogica(FILE* disk, long long startExtendida, std::string nombre);
+    void printLogicas(FILE* disk, long long startExtendida);
     void execute(int size, char unit, std::string path,
                  char type, std::string fit, std::string name);
 };
-
 #endif
